move batch, sort_embeddings out of backends/candle/tests/
