+++ backends/candle/src/lib.rs
@@ -24,11 +24,12 @@ use candle::{DType, Device};
 use candle_nn::VarBuilder;
 use nohash_hasher::BuildNoHashHasher;
 use serde::Deserialize;
-use std::collections::HashMap;
 use std::path::PathBuf;
+use std::{cmp::max, collections::HashMap};
 use text_embeddings_backend_core::{
     Backend, BackendError, Batch, Embedding, Embeddings, ModelType, Predictions,
 };
+use tokenizers::Encoding;
 
 /// This enum is needed to be able to differentiate between jina models that also use
 /// the `bert` model type and valid Bert models.
@@ -372,4 +373,49 @@ impl<O> WrapErr<O> for Result<O, candle::Error> {
     fn e(self) -> Result<O, BackendError> {
         self.map_err(|e| BackendError::Inference(e.to_string()))
     }
+}
+
+pub fn batch(encodings: Vec<Encoding>, pooled_indices: Vec<u32>, raw_indices: Vec<u32>) -> Batch {
+    let mut input_ids = Vec::new();
+    let mut token_type_ids = Vec::new();
+    let mut position_ids = Vec::new();
+    let mut cumulative_seq_lengths = Vec::with_capacity(encodings.len() + 1);
+    cumulative_seq_lengths.push(0);
+
+    let mut max_length = 0;
+    let mut cumulative_length = 0;
+
+    for encoding in encodings.iter() {
+        let encoding_length = encoding.len() as u32;
+        input_ids.extend(encoding.get_ids().to_vec());
+        token_type_ids.extend(encoding.get_type_ids().to_vec());
+        position_ids.extend(0..encoding_length);
+        cumulative_length += encoding_length;
+        cumulative_seq_lengths.push(cumulative_length);
+        max_length = max(max_length, encoding_length);
+    }
+
+    Batch {
+        input_ids,
+        token_type_ids,
+        position_ids,
+        cumulative_seq_lengths,
+        max_length,
+        pooled_indices,
+        raw_indices,
+    }
+}
+
+pub fn sort_embeddings(embeddings: Embeddings) -> (Vec<Vec<f32>>, Vec<Vec<f32>>) {
+    let mut pooled_embeddings = Vec::new();
+    let mut raw_embeddings = Vec::new();
+
+    for (_, embedding) in embeddings {
+        match embedding {
+            Embedding::Pooled(e) => pooled_embeddings.push(e),
+            Embedding::All(e) => raw_embeddings.extend(e),
+        }
+    }
+
+    (pooled_embeddings, raw_embeddings)
 }
