+++ backends/candle/tests/test_flash_nomic.rs
@@ -4,7 +4,7 @@ mod common;
 use crate::common::{sort_embeddings, SnapshotScores};
 use anyhow::Result;
 use common::{batch, download_artifacts, load_tokenizer, relative_matcher};
-use text_embeddings_backend_candle::CandleBackend;
+use text_embeddings_backend_candle::{batch, sort_embeddings, CandleBackend};
 use text_embeddings_backend_core::{Backend, ModelType, Pool};
 
 #[test]
